requirements.txt
faiss-cpu==1.10.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-google-genai==2.0.10
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
openai==1.68.2
pypdf==5.4.0
python-dotenv==1.0.1
streamlit==1.43.2

